Added thpool_thread_index() to the API, plus change to thpool_init().
One case where our application wanted to violate the "opaque handles" of
the previous commit, was to associate a pthread_t with the internal index
in the array of threadpool threads.  This is useful to allow an application
to have its own arrays of data that have per-thread elements.  However, We
don't want to provide an extension to the API that allows callers to see
the pthread_t values in its internal arrays.

To address this, `thpool_thread_index()` requires the caller to provide a
pthread_t, and returns the corresponding index.  Thus, we don't expose any
thread-handles to the caller that they don't already have.

Also, in the previous commit thpool_init() was still returning NULL instead
of the new opaque handles, in the case of errors.  It now returns a handle
containing a NULL pointer.  All existing apps that test the return value
will have to be changed.  We provide `thpool_null()` for this.  Such apps
should now use something like this:

   if (thpool_null(threadpool))
      // thpool_init() failed ...